# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Стильный призыв API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /posts:
    get:
      operationId: listPosts
      tags: [posts]
      summary: List published posts
      responses:
        "200":
          description: Posts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListPostsPayload"
    post:
      operationId: createPost
      tags: [posts]
      summary: Create a post (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreatePostPayload"
...
  /posts/{slug}:
    get:
      operationId: getPostBySlug
      tags: [posts]
      summary: Get a published post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetPostBySlugPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
  /posts/id/{id}:
    patch:
      operationId: updatePost
      tags: [posts]
      summary: Update a post (owner, or editor for own post)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdatePostPayload"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    delete:
      operationId: deletePost
      tags: [posts]
      summary: Delete a post (owner, or editor for own post)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Post interactions ──────────────────────────────────────────────────
...
  /hadiths:
    get:
      operationId: listHadiths
      tags: [hadiths]
      summary: List published hadiths
      parameters:
        - name: grade
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/HadithGrade"
        - name: topic
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Hadiths
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListHadithsPayload"
    post:
      operationId: createHadith
      tags: [hadiths]
      summary: Create a hadith (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HadithInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateHadithPayload"
...
  /hadiths/{id}:
    get:
      operationId: getHadith
      tags: [hadiths]
      summary: Get a single published hadith
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Hadith
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetHadithPayload"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    patch:
      operationId: updateHadith
      tags: [hadiths]
      summary: Update a hadith (owner, or editor for own hadith)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HadithUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateHadithPayload"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"
    delete:
      operationId: deleteHadith
      tags: [hadiths]
      summary: Delete a hadith (owner, or editor for own hadith)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorPayload"

  # ── Admins ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorPayload:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Post:
      type: object
      required: [id, title, slug, excerpt, content, published, authorName, createdAt, updatedAt]
      properties:
        id: { type: integer }
        title: { type: string }
        slug: { type: string }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        published: { type: boolean }
        authorId: { type: ["integer", "null"] }
        authorName: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ListPostsPayload:
      type: array
      items:
        $ref: "#/components/schemas/Post"
...
    PostInput:
      type: object
      required: [title]
      properties:
        title: { type: string, minLength: 1 }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        published: { type: boolean }
    PostUpdate:
      type: object
      properties:
        title: { type: string, minLength: 1 }
        excerpt: { type: string }
        content: { type: string }
        coverImageUrl: { type: ["string", "null"] }
        published: { type: boolean }
    CreatePostPayload:
      $ref: "#/components/schemas/Post"
    GetPostBySlugPayload:
      $ref: "#/components/schemas/Post"
    UpdatePostPayload:
      $ref: "#/components/schemas/Post"

    # ── Post interactions ──────────────────────────────────────────────
...
    HadithGrade:
      type: string
      enum: [sahih, hasan, daif, mawdu]
    Hadith:
      type: object
      required: [id, text, source, narrator, grade, topic, published, authorName, createdAt, updatedAt]
      properties:
        id: { type: integer }
        text: { type: string }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
        authorId: { type: ["integer", "null"] }
        authorName: { type: string }
        createdAt: { type: string }
        updatedAt: { type: string }
    ListHadithsPayload:
      type: array
      items:
        $ref: "#/components/schemas/Hadith"
...
    HadithInput:
      type: object
      required: [text, grade]
      properties:
        text: { type: string, minLength: 1 }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
    HadithUpdate:
      type: object
      properties:
        text: { type: string, minLength: 1 }
        source: { type: string }
        narrator: { type: string }
        grade:
          $ref: "#/components/schemas/HadithGrade"
        topic: { type: string }
        published: { type: boolean }
    GetHadithPayload:
      $ref: "#/components/schemas/Hadith"
    CreateHadithPayload:
      $ref: "#/components/schemas/Hadith"
    UpdateHadithPayload:
      $ref: "#/components/schemas/Hadith"

    # ── Messages ───────────────────────────────────────────────────────
...