# lib/api-zod/src/generated/types/post.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Стильный призыв API
 * OpenAPI spec version: 0.1.0
 */

export interface Post {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  coverImageUrl?: string | null;
  published: boolean;
  /** @nullable */
  authorId?: number | null;
  authorName: string;
  createdAt: string;
  updatedAt: string;
}