# lib/api-zod/src/generated/types/postUpdate.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Стильный призыв API
 * OpenAPI spec version: 0.1.0
 */

export interface PostUpdate {
  /** @minLength 1 */
  title?: string;
  excerpt?: string;
  content?: string;
  /** @nullable */
  coverImageUrl?: string | null;
  published?: boolean;
}